Changed Some Features on the keypad class
Put all the keypad buttons on, but I didn't implement an array. It is
dynamic so when you instantiate the keypad class and you stipulate the
bounds the keypad goes a third of the way to the right and is 3/4 the
height of the window and the buttons all wrap according to the bounds of
the panel, try it out... just instanitate the keypad object. ie JPanel
keypad = new KeyPad(100,100,100,100).